requirements.txt
Django==4.0
mysqlclient
channels
djangorestframework
django-filter
django-environ
requests
